fix: parseInt→parseFloat for decimal encounter weights; make roll test deterministic
- roll_encounter was using parseInt() on weights like 0.99/0.01, truncating
  both to 0 and making selection 50/50 instead of respecting declared weights
- The Workers runtime re-seeds Math.random() on reset(), making repeated
  probabilistic assertions flaky; replaced count check with membership check
  (selected_archetype must be one of the two seeded archetypes)

+++ src/tools/entity.ts
@@ -471,7 +471,7 @@ export async function handle_roll_encounter({ c, id, args }: ToolContext): Promi
   const entries: Array<{ key: string; weight: number }> = []
   for (const part of tableRaw.split(',').map(s => s.trim()).filter(Boolean)) {
     const m = part.match(/^(.+?)\s*:\s*([\d.]+)$/)
-    if (m) entries.push({ key: m[1].trim(), weight: parseInt(m[2]) })
+    if (m) entries.push({ key: m[1].trim(), weight: parseFloat(m[2]) })
     else entries.push({ key: part, weight: 1 })
   }
   if (entries.length === 0) {
